split out utils for making constants out of values
PiperOrigin-RevId: 807683774
